requirements.txt
amqp==2.6.0
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-admin==2.0.1
django-allauth==0.42.0
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-six==1.0.4
django-widget-tweaks==1.4.8
easy-thumbnails==2.7
future==0.18.2
idna==2.10
kombu==4.6.11
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
screen==1.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
xlwt==1.3.0
